# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/application/accept/{application}": {
            "get": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Accept an application",
                "description": "Accept an application.",
                "operationId": "1ea4cb387ee9d0a5134d8a88ff513466",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID of the application to accept",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application accepted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Application accepted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Application not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/application/reject/{application}": {
            "get": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Reject an application",
                "description": "Reject an application.",
                "operationId": "0cadec3d6ab7757274d94b7cbf8cd86c",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID of the application to reject",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application rejected successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Application rejected successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Application not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/applications/requests/all": {
            "get": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Retrieve all pending applications",
                "description": "Retrieve all pending applications.",
                "operationId": "925d92a90d8bf66e1d0716a0d9b807e9",
                "responses": {
                    "200": {
                        "description": "List of all pending applications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "pending_applications": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/application/create": {
            "post": {
                "tags": [
                    "Volunteer"
                ],
                "summary": "Apply for an announcement",
                "description": "Apply for an announcement.",
                "operationId": "6573aa85cf203124c6e1de9b1b71b58b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "announcement_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Application done successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Application done successfully, waiting for admin approval"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    }
}